chore(desktop): remove related code of market banner ff
