+++ apps/ledger-live-desktop/src/mvvm/components/SideBar/types.ts
@@ -34,7 +34,6 @@ export interface SideBarViewModel {
   readonly isCardDisabled: boolean;
   readonly isAccountsDisabled: boolean;
   readonly isLiveAppTabSelected: boolean;
-  readonly isMarketBannerEnabled: boolean;
   readonly isWallet40MainNavEnabled: boolean;
   readonly isMyWalletEnabled: boolean;
   readonly referralProgramConfig: ReferralProgramConfig | null;
