+++ apps/ledger-live-desktop/src/mvvm/features/Portfolio/PortfolioView.tsx
@@ -25,7 +25,6 @@ export const PortfolioView = memo(function PortfolioView({
   totalOperations,
   totalCurrencies,
   hasExchangeBannerCTA,
-  shouldDisplayMarketBanner,
   shouldDisplayGraphRework,
   shouldDisplayAssetSection,
   shouldDisplayAssetDiscoverability,
@@ -67,7 +66,8 @@ export const PortfolioView = memo(function PortfolioView({
           </div>
 
           <PortfolioBannerContent />
-          {shouldDisplayMarketBanner && <MarketBanner />}
+
+          <MarketBanner />
 
           <PerpsEntryPoint />
