+++ apps/ledger-live-desktop/src/mvvm/features/Portfolio/__integrations__/Portfolio.integration.test.tsx
@@ -165,7 +165,6 @@ describe("PortfolioView", () => {
     totalOperations: 10,
     totalCurrencies: 3,
     hasExchangeBannerCTA: true,
-    shouldDisplayMarketBanner: true,
     shouldDisplayGraphRework: true,
     shouldDisplayAssetSection: true,
     shouldDisplayAssetDiscoverability: false,
@@ -434,7 +433,7 @@ describe("PortfolioView", () => {
         }),
       );
 
-      render(<PortfolioView {...defaultProps} shouldDisplayMarketBanner={true} />);
+      render(<PortfolioView {...defaultProps} />);
 
       await waitFor(() => {
         expect(screen.getByTestId("trending-assets-list")).toBeVisible();
@@ -451,7 +450,7 @@ describe("PortfolioView", () => {
         }),
       );
 
-      render(<PortfolioView {...defaultProps} shouldDisplayMarketBanner={true} />);
+      render(<PortfolioView {...defaultProps} />);
 
       expect(screen.getByTestId("skeleton-list")).toBeVisible();
     });
@@ -463,16 +462,11 @@ describe("PortfolioView", () => {
         }),
       );
 
-      render(<PortfolioView {...defaultProps} shouldDisplayMarketBanner={true} />);
+      render(<PortfolioView {...defaultProps} />);
 
       await waitFor(() => {
         expect(screen.getByTestId("generic-error")).toBeVisible();
       });
-    });
-
-    it("should not render MarketBanner when shouldDisplayMarketBanner is false", () => {
-      render(<PortfolioView {...defaultProps} shouldDisplayMarketBanner={false} />);
-      expect(screen.queryByText("Market")).toBeNull();
     });
   });
