+++ apps/ledger-live-desktop/src/renderer/Default.tsx
@@ -89,8 +89,7 @@ const SwapWeb = lazy(() => import("~/renderer/screens/swapWeb"));
 const Swap2 = lazy(() => import("~/renderer/screens/exchange/Swap2"));
 const Perps = lazy(() => import("LLD/features/Perps"));
 const Borrow = lazy(() => import("LLD/features/Borrow"));
-const Market40 = lazy(() => import("LLD/features/Market"));
-const Market = lazy(() => import("~/renderer/screens/market"));
+const Market = lazy(() => import("LLD/features/Market"));
 
 const MarketCoin = lazy(() => import("~/renderer/screens/market/MarketCoin"));
 const WelcomeScreenSettings = lazy(
@@ -254,12 +253,10 @@ const RecoverPlayerWithFeatureToggle = () => {
 
 // Shared content for the main app layout
 const MainAppContent = ({
-  shouldDisplayMarketBanner,
   shouldDisplayWallet40MainNav,
   shouldDisplayAssetSection,
   shouldDisplayAggregatedAssets,
 }: {
-  shouldDisplayMarketBanner: boolean;
   shouldDisplayWallet40MainNav: boolean;
   shouldDisplayAssetSection: boolean;
   shouldDisplayAggregatedAssets: boolean;
@@ -324,10 +321,7 @@ const MainAppContent = ({
             shouldDisplayAggregatedAssets ? <RedirectMarketToAsset /> : withSuspense(MarketCoin)({})
           }
         />
-        <Route
-          path="/market"
-          element={withSuspense(shouldDisplayMarketBanner ? Market40 : Market)({})}
-        />
+        <Route path="/market" element={withSuspense(Market)({})} />
         <Route path="/bank/*" element={withSuspense(Bank)({})} />
         <Route path="/analytics" element={withSuspense(Analytics)({})} />
         <Route path="/history" element={withSuspense(History)({})} />
@@ -344,7 +338,6 @@ export const MainAppLayout = () => {
   const theme = useSelector(themeSelector);
   const styledComponentsTheme = useTheme();
   const {
-    shouldDisplayMarketBanner,
     isEnabled: isWallet40Enabled,
     shouldDisplayWallet40MainNav,
     shouldDisplayAssetSection,
@@ -393,7 +386,6 @@ export const MainAppLayout = () => {
         }
       >
         <MainAppContent
-          shouldDisplayMarketBanner={shouldDisplayMarketBanner}
           shouldDisplayWallet40MainNav={shouldDisplayWallet40MainNav}
           shouldDisplayAssetSection={shouldDisplayAssetSection}
           shouldDisplayAggregatedAssets={shouldDisplayAggregatedAssets}
