+++ apps/ledger-live-desktop/src/renderer/screens/market/components/MarketItemChart.tsx
