+++ apps/ledger-live-desktop/src/renderer/screens/market/components/SideDrawerFilter.tsx
@@ -2,7 +2,6 @@ import React, { useCallback } from "react";
 import { TFunction } from "i18next";
 import Dropdown from "./DropDown";
 import { MarketListRequestParams } from "@ledgerhq/live-common/market/utils/types";
-import { useWalletFeaturesConfig } from "@features/platform-feature-flags";
 
 export default function SideDrawerFilter({
   refresh,
@@ -18,14 +17,7 @@ export default function SideDrawerFilter({
 }) {
   const { starred, liveCompatible } = filters;
 
-  const { shouldDisplayMarketBanner: filterBySupported } = useWalletFeaturesConfig("desktop");
-
-  const shouldDisplayCompatibleOption = !filterBySupported;
-
-  const resetFilters = useCallback(
-    () => refresh({ starred: [], liveCompatible: filterBySupported }),
-    [refresh, filterBySupported],
-  );
+  const resetFilters = useCallback(() => refresh({ starred: [], liveCompatible: true }), [refresh]);
   const onChange = useCallback(
     (option?: { label: string; value: string } | null) => {
       if (!option) return;
@@ -49,14 +41,6 @@ export default function SideDrawerFilter({
       value: "all",
       label: t("market.filters.all"),
     },
-    ...(shouldDisplayCompatibleOption
-      ? [
-          {
-            value: "liveCompatible",
-            label: t("market.filters.isLedgerCompatible"),
-          },
-        ]
-      : []),
     {
       value: "starred",
       label: t("market.filters.isFavorite"),
@@ -78,14 +62,6 @@ export default function SideDrawerFilter({
                 {
                   value: "all",
                   label: t("market.filters.all"),
-                },
-              ]
-            : []),
-          ...(shouldDisplayCompatibleOption && liveCompatible.value
-            ? [
-                {
-                  value: "liveCompatible",
-                  label: t("market.filters.isLedgerCompatible"),
                 },
               ]
             : []),
